# docs/10-critical-remediation/api-reference-openapi.yaml
openapi: 3.1.0
info:
  title: Ubertruck MVP API Reference
  description: Complete REST API specification for Ubertruck Logistics Platform
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@ubertruck.in
  license:
    name: Proprietary
    url: https://ubertruck.in/terms
servers:
  - url: https://api.ubertruck.in/v1
    description: Production Server
  - url: https://staging-api.ubertruck.in/v1
    description: Staging Server
  - url: http://localhost:3000/v1
    description: Development Server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
        - timestamp
        - requestId
      properties:
        code:
          type: string
          pattern: '^[A-Z]{3}_[A-Z_]+$'
          example: 'USR_NOT_FOUND'
        message:
          type: string
          example: 'User not found with the provided ID'
        details:
          type: object
          properties:
            field:
              type: string
            reason:
              type: string
            suggestion:
              type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
        downstream:
          type: object
          properties:
            service:
              type: string
            error:
              type: string
            statusCode:
              type: integer
...
    OTPVerification:
      type: object
      required:
        - phoneNumber
        - otp
      properties:
        phoneNumber:
          type: string
          pattern: '^\+91[6-9]\d{9}$'
        otp:
          type: string
          pattern: '^\d{6}$'
        sessionId:
          type: string
          format: uuid
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          example: 3600
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            phoneNumber:
              type: string
            businessName:
              type: string
            userType:
              type: string
            verified:
              type: boolean

    # Fleet Management Schemas
...
paths:
...
  /auth/verify-otp:
    post:
      tags:
        - Authentication
      summary: Verify OTP and complete registration
      operationId: verifyOTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OTPVerification'
      responses:
        '200':
          description: OTP verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many failed attempts
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until retry allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /tracking/{bookingId}/pod:
    post:
      tags:
        - Tracking
      summary: Upload Proof of Delivery
      operationId: uploadPOD
      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - podImage
              properties:
                podImage:
                  type: string
                  format: binary
                  description: POD image (max 2MB, JPEG/PNG)
                receiverName:
                  type: string
                receiverSignature:
                  type: string
                  format: byte
                  description: Base64 encoded signature
      responses:
        '201':
          description: POD uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  podUrl:
                    type: string
                    format: uri
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Invalid file type or booking status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Payment Endpoints
...